@@ -20,31 +20,31 @@
   },
   "dependencies": {
     "@beak/design-system": "*",
-    "@types/node": "^20.5.9",
+    "@types/node": "^20.10.5",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
-    "eslint": "^8.48.0",
+    "eslint": "^8.56.0",
     "eslint-config-next": "^13.4.19",
-    "next": "^13.4.19",
+    "next": "^14.0.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.3"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.4.2",
-    "@fortawesome/free-brands-svg-icons": "^6.4.2",
-    "@fortawesome/free-regular-svg-icons": "^6.4.2",
-    "@fortawesome/free-solid-svg-icons": "^6.4.2",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@types/js-cookie": "^3.0.3",
-    "@types/js-yaml": "^4.0.5",
-    "@types/react-helmet": "^6.1.6",
-    "@types/react-redux": "^7.1.26",
+    "@reduxjs/toolkit": "^2.0.1",
+    "@types/js-cookie": "^3.0.6",
+    "@types/js-yaml": "^4.0.9",
+    "@types/react-helmet": "^6.1.11",
+    "@types/react-redux": "^7.1.33",
     "js-cookie": "^3.0.5",
     "js-yaml": "^4.1.0",
     "react-helmet": "^6.1.0",
-    "react-redux": "^8.1.2",
+    "react-redux": "^9.0.4",
     "redux-devtools-extension": "^2.13.9"
   }
 }
