@@ -52,55 +52,55 @@
     "release-safety": "yarn typecheck && yarn lint && yarn build"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.22.14",
-    "@babel/preset-react": "^7.22.5",
-    "@babel/preset-typescript": "^7.22.11",
+    "@babel/preset-env": "^7.23.6",
+    "@babel/preset-react": "^7.23.3",
+    "@babel/preset-typescript": "^7.23.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/node": "^20.5.9",
+    "@testing-library/user-event": "^14.5.1",
+    "@types/node": "^20.10.5",
     "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
-    "@types/sanitize-html": "^2.9.0",
-    "@types/semver": "^7.5.1",
-    "@types/url-parse": "^1.4.9",
-    "@typescript-eslint/eslint-plugin": "^6.5.0",
-    "@typescript-eslint/parser": "^6.5.0",
-    "@vitejs/plugin-react": "^4.0.4",
+    "@types/sanitize-html": "^2.9.5",
+    "@types/semver": "^7.5.6",
+    "@types/url-parse": "^1.4.11",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "^6.15.0",
+    "@vitejs/plugin-react": "^4.2.1",
     "esbuild-jest": "^0.5.0",
-    "eslint": "^8.48.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-jest": "^26.8.2",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-unused-imports": "^3.0.0",
-    "lerna": "^7.2.0",
-    "nx": "^16.7.4",
-    "react-spring": "^9.7.2",
+    "lerna": "^8.0.1",
+    "nx": "^17.2.7",
+    "react-spring": "^9.7.3",
     "sanitize-html": "^2.11.0",
     "semver": "^7.5.4",
     "ts-md5": "^1.3.1",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.10",
     "vite-plugin-sentry": "^1.3.0"
   },
   "dependencies": {
-    "@sentry/cli": "^2.20.6",
-    "@sentry/electron": "^4.10.0",
-    "@sentry/react": "^7.66.0",
-    "@sentry/tracing": "^7.66.0",
+    "@sentry/cli": "^2.23.1",
+    "@sentry/electron": "^4.15.1",
+    "@sentry/react": "^7.91.0",
+    "@sentry/tracing": "^7.91.0",
     "chokidar": "^3.5.3",
     "cross-env": "^7.0.3",
     "hw-fingerprint": "^2.1.0",
     "jest": "^27.5.1",
     "keytar": "^7.9.0",
-    "prettier": "^3.0.3",
+    "prettier": "^3.1.1",
     "pretty-bytes": "^6.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router": "^6.15.0",
-    "react-router-dom": "^6.15.0",
+    "react-router": "^6.21.1",
+    "react-router-dom": "^6.21.1",
     "tslog": "^4.9.2",
     "url-parse": "^1.5.10",
     "valid-filename": "^4.0.0",
