@@ -1,13 +1,13 @@
 /* eslint-disable no-process-env */
 /* eslint-disable @typescript-eslint/no-var-requires */
 
-const path = require('path');
-const fs = require('fs');
-const reactPlugin = require('@vitejs/plugin-react');
-const viteSentryPlugin = require('vite-plugin-sentry');
-const monacoEditorPlugin = require('vite-plugin-monaco-editor');
-const mkcert = require('vite-plugin-mkcert');
-const { NodeGlobalsPolyfillPlugin } = require('@esbuild-plugins/node-globals-polyfill');
+import { NodeGlobalsPolyfillPlugin } from '@esbuild-plugins/node-globals-polyfill';
+import reactPlugin from '@vitejs/plugin-react';
+import fs from 'fs';
+import path from 'path';
+import mkcert from 'vite-plugin-mkcert';
+import monacoEditorPlugin from 'vite-plugin-monaco-editor';
+import viteSentryPlugin from 'vite-plugin-sentry';
 
 // eslint-disable-next-line no-sync
 const packageJson = JSON.parse(fs.readFileSync(path.join(
@@ -29,7 +29,7 @@ const sourcePathInDev = environment === 'development' ? 'src' : '';
  * @type {import('vite').UserConfig}
  * @see https://vitejs.dev/config/
  */
-module.exports = {
+export default {
 	mode: environment,
 	jsx: 'react',
 	root: './src',
@@ -53,7 +53,7 @@ module.exports = {
 		https: true,
 	},
 	plugins: [
-		mkcert.default(),
+		mkcert(),
 		reactPlugin({ include: '**/*.tsx' }),
 		monacoEditorPlugin.default({
 			globalAPI: true,
